ocrapp/requirements.txt
fastapi==0.95.2
uvicorn==0.23.1
sqlmodel==0.0.8
SQLAlchemy==1.4.32
pytesseract==0.3.9
Pillow==9.2.0
python-docx==0.8.11
